pre-check warning for max_completions_token

+++ pkg/model_config.go
@@ -5,11 +5,11 @@ import (
 )
 
 type ModelConfigT struct {
-	Id                      string        `mapstructure:"id,omitempty"`
-	Name                    string        `mapstructure:"name,omitempty"`
-	Temperature             float64       `mapstructure:"temperature,omitempty"`
-	MaxCompletionTokens     int64         `mapstructure:"max_completion_tokens"`
-	ContextWindow           int           `mapstructure:"context_window"`
+	Id                  string  `mapstructure:"id,omitempty"`
+	Name                string  `mapstructure:"name,omitempty"`
+	Temperature         float64 `mapstructure:"temperature,omitempty"`
+	MaxCompletionTokens int64   `mapstructure:"max_completion_tokens"`
+	ContextWindow       int64   `mapstructure:"context_window"`
 	// TikTokenEnabled         bool          `mapstructure:"tik_token_enabled"`
 	ApiKey                  string        `mapstructure:"api_key"`
 	BaseUrl                 string        `mapstructure:"base_url"`
